+++ src/pages/RootPage.jsx
@@ -1,13 +1,13 @@
 import MadeWithLove from '@/components/welcome/MadeWithLove';
+import ProfilePopover from '@/components/welcome/ProfilePopover';
 import WelcomePanel from '@/components/welcome/WelcomePanel';
-import AutoInstallPopover from '@/progressive/AutoInstallPopover';
 
 export default function RootPage() {
   return (
     <main className="w-full h-full flex flex-col items-center bg-white">
       <WelcomePanel />
       <MadeWithLove />
-      <AutoInstallPopover />
+      <ProfilePopover />
     </main>
   );
 }